+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ------------------
 
 - fixed: tooltips (popover) position in panel.
+- fixed: time slideshow rounding number of steps and rounding pb
+- added: panel and site config yaml parsing errors checks and display.
+- reordered: panel options display data time controls
 
 ## 1.20.1